// src/__generated__/checkUserName.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: checkUserName
// ====================================================

export interface checkUserName {
  checkUserName: boolean;
}

export interface checkUserNameVariables {
  userName: string;
}